This adds prettier and linting
We like to use prettier for formatting to keep a consistent style in the
code and want to maintain a clean linting as well. This adds prettier,
linting and formats the code to clean everything up.

Upgrade some packages such as the zeal scripts.

Also removes redux and store concepts as this is a simple site.